+++ src/Sonic/Protocol.hs
@@ -1,44 +1,54 @@
 -- The interactive Sonic protocol to check that the prover knows a valid assignment of the wires in the circuit
 
 {-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DeriveAnyClass #-}
 module Sonic.Protocol
   ( Proof
+  , RndOracle(..)
   , prove
   , verify
   ) where
 
 import Protolude hiding (head)
 import Data.List (head)
+import Data.Pairing.BLS12381 (Fr, G1, BLS12381)
 import Control.Monad.Random (MonadRandom)
 import Bulletproofs.ArithmeticCircuit (ArithCircuit(..), Assignment(..), GateWeights(..))
 import Math.Polynomial.Laurent (newLaurent, evalLaurent)
-import GaloisField (GaloisField(rnd))
+import Data.Field.Galois (rnd)
 
 import Sonic.SRS (SRS(..))
 import Sonic.Constraints (rPoly, sPoly, tPoly, kPoly)
 import Sonic.CommitmentScheme (commitPoly, openPoly, pcV)
 import Sonic.Signature (HscProof(..), hscP, hscV)
 import Sonic.Utils (evalOnY)
-import Sonic.Curve (Fr, G1)
 
 data Proof f = Proof
-  { prR :: G1
-  , prT :: G1
+  { prR :: G1 BLS12381
+  , prT :: G1 BLS12381
   , prA :: f
-  , prWa :: G1
+  , prWa :: G1 BLS12381
   , prB :: f
-  , prWb :: G1
-  , prWt :: G1
+  , prWb :: G1 BLS12381
+  , prWt :: G1 BLS12381
   , prS :: f
   , prHscProof :: HscProof f
-  }
+  } deriving (Eq, Show, Generic, NFData)
+
+-- | Values created non-interactively in the random oracle model during proof generation
+data RndOracle f = RndOracle
+  { rndOracleY :: f
+  , rndOracleZ :: f
+  , rndOracleYs :: [f]
+  } deriving (Eq, Show, Generic, NFData)
 
 prove
   :: MonadRandom m
   => SRS
   -> Assignment Fr
   -> ArithCircuit Fr
-  -> m (Proof Fr, Fr, Fr, [Fr])
+  -> m (Proof Fr, RndOracle Fr)
 prove srs@SRS{..} assignment@Assignment{..} arithCircuit@ArithCircuit{..} =
   if srsD < 7*n
     then panic $ "Parameter d is not large enough: " <> show srsD <> " should be greater than " <>  show (7*n)
@@ -80,9 +90,11 @@ prove srs@SRS{..} assignment@Assignment{..} arithCircuit@ArithCircuit{..} =
            , prS = s
            , prHscProof = hscProof
            }
-         , y
-         , z
-         , ys
+         , RndOracle
+           { rndOracleY = y
+           , rndOracleZ = z
+           , rndOracleYs = ys
+           }
          )
   where
     n :: Int